refactor: NotificationType Enum을 Notification 도메인 아래로 변경
