+++ src/main/java/umc/cockple/demo/domain/notification/domain/Notification.java
@@ -4,7 +4,7 @@ import jakarta.persistence.*;
 import lombok.*;
 import umc.cockple.demo.domain.member.domain.Member;
 import umc.cockple.demo.global.common.BaseEntity;
-import umc.cockple.demo.global.enums.NotificationType;
+import umc.cockple.demo.domain.notification.enums.NotificationType;
 
 @Entity
 @Builder
