+++ src/main/java/umc/cockple/demo/domain/notification/dto/MarkAsReadDTO.java
@@ -1,6 +1,6 @@
 package umc.cockple.demo.domain.notification.dto;
 
-import umc.cockple.demo.global.enums.NotificationType;
+import umc.cockple.demo.domain.notification.enums.NotificationType;
 
 public class MarkAsReadDTO {
